Emit `IKeyboardLayoutService.onDidChangeKeyboardLayout` only in case of a change
